% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcIncomeTax.R
\name{calcIncomeTax}
\alias{calcIncomeTax}
\title{Calculate income after tax and benefits}
\usage{
calcIncomeTax(borrowers, stress = FALSE, stress.args = list(inflation = 1,
  years = 3, childBenefitChange = 1, personalAllowanceChange = 500),
  personalAllowanceThreshold = 1e+05, personalAllowanceValue = 10000,
  studentLoanThreshold = 16910, studentLoanPercentage = 0.09,
  childBenefitThreshold = 50000, childBenefitChild1 = 1066,
  childBenefitChild2 = 704.6, taxBrackets = fread(system.file("extdata",
  "taxrates.csv", package = "optiRum")),
  class1NIBrackets = fread(system.file("extdata",
  "NationalInsuranceThresholds.csv", package = "optiRum")),
  class4NIBrackets = fread(system.file("extdata",
  "NationalInsurance4Thresholds.csv", package = "optiRum")))
}
\arguments{
\item{borrowers}{data.table with following columns \emph{borrowerId, loanId, employedIncome, investmentIncome, nonTaxableIncome, selfEmployedProfits, taxCode, numberOfChildren, salarySacrificePercentage, studentLoan}.}

\item{stress}{logical}

\item{stress.args}{list of arguments used in stress income/expenditure: \emph{inflation, years, childBenefitChange, personalAllowanceChange}.}

\item{personalAllowanceThreshold}{numeric, default 100000.}

\item{personalAllowanceValue}{numeric, default 10000.}

\item{studentLoanThreshold}{numeric, default 16910.}

\item{studentLoanPercentage}{numeric, default 0.09.}

\item{childBenefitThreshold}{numeric, default 50000.}

\item{childBenefitChild1}{numeric, default 1066.}

\item{childBenefitChild2}{numeric, default 704.6.}

\item{taxBrackets}{data.table.}

\item{class1NIBrackets}{data.table.}

\item{class4NIBrackets}{data.table.}
}
\description{
Calculate income after tax and benefits
}

