\name{logit.probs}
\alias{logit.probs}
\title{Convert a logit to probability}
\usage{
logit.probs(logit)
}
\arguments{
  \item{logit}{The log(odds)}
}
\value{
prob Probability
}
\description{
Transform a logit response from a glm into probability
}
\seealso{
glm
}
\keyword{glm}
\keyword{logit}
\keyword{probability}

