\name{odd.prob}
\alias{odd.prob}
\title{Convert an odds to probability}
\usage{
odd.prob(odds)
}
\arguments{
  \item{odds}{Odds}
}
\value{
prob Probability
}
\description{
Transform odds into a probability
}
\examples{
odd.prob(1) # equals 0.5
}
\seealso{
\code{\link{logit.odd}} \code{\link{logit.prob}}

Other creditrisk: \code{\link{giniChart}};
\code{\link{giniCoef}}; \code{\link{logit.odd}};
\code{\link{logit.prob}}; \code{\link{odd.logit}};
\code{\link{prob.logit}}; \code{\link{prob.odd}};
\code{\link{scaledScore}}
}
\keyword{glm}
\keyword{logit}
\keyword{odds}
\keyword{probability}

