% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optiRum-package.r
\docType{package}
\name{optiRum}
\alias{optiRum}
\title{optiRum is a helper package}
\description{
optiRum is a growing package of utilities created by Optimum Credit Ltd's analysts.  
It is designed to provide convenience functions, standards, and useful snippets.  
Optimum Credit derives significant value from the R platform and associated community, 
so non-commercially sensitive functionality is made available in the spirit of reciprocity.
}
