\name{prob.logit}
\alias{prob.logit}
\title{Convert a probability into a logit}
\usage{
prob.logit(prob)
}
\arguments{
  \item{prob}{Probability}
}
\value{
logit Log(odds)
}
\description{
Transforming probabilities into logits (the response from
binomial glms)
}
\keyword{glm}
\keyword{logit}
\keyword{odds}
\keyword{probability}

