% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise.R
\name{sanitise}
\alias{sanitise}
\title{A cleaning function for special characters}
\usage{
sanitise(str)
}
\arguments{
\item{str}{The text to be sanitised}
}
\description{
This function is a helper for cleaning xtable outputs in preparation for PDF production
}
\examples{
sanitise('[&\%#<>\\\\')

}
\seealso{
Other helper: 
\code{\link{CJ.dt}()},
\code{\link{convertToXML}()},
\code{\link{generatePDF}()},
\code{\link{pounds_format}()},
\code{\link{thousands_format}()},
\code{\link{wordwrap}()}
}
\concept{helper}
\keyword{PDF}
\keyword{Rnw}
\keyword{generate}
\keyword{knitr}
\keyword{pdflatex}
