\name{theme_optimum}
\alias{theme_optimum}
\title{Produce an Optimum-standard base chart}
\usage{
theme_optimum(base_size = 14, base_family = "Lato Light", ...)
}
\arguments{
  \item{base_size}{Anchor font size}

  \item{base_family}{Font family to use}

  \item{...}{Allows additional parameters to be passed to
  further themes}
}
\description{
This theme builds on the Stephen Few theme from ggthemes to
produce a good baseline for charting in R. Gets called as
would any typical theme.
}
\examples{

}
\keyword{Few}
\keyword{ggplot2}
\keyword{theme}

