% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thousands.R
\name{thousands_format}
\alias{thousands_format}
\alias{thousands}
\title{Thousands formatter: format number with commas separating the number thousands and suffixed with a k.}
\usage{
thousands_format()

thousands(x)
}
\arguments{
\item{x}{a numeric vector to format}
}
\value{
a function with single paramater x, a numeric vector, that
  returns a character vector
}
\description{
Based heavily on the scales work by Hadley
}
\examples{
thousands_format()(c(1, 1e3, 2000, 1e6))
thousands_format()(c(1, 1e3, 2000, 1e6))
thousands(c(1, 1e3, 2000, 1e6))
}
\seealso{
Other helper: 
\code{\link{CJ.dt}()},
\code{\link{convertToXML}()},
\code{\link{generatePDF}()},
\code{\link{pounds_format}()},
\code{\link{sanitise}()},
\code{\link{wordwrap}()}
}
\concept{helper}
