\docType{package}
\name{optiRum}
\alias{optiRum}
\alias{optiRum-package}
\title{optiRum is a helper package}
\description{
optiRum is a growing package of utilities created by
Optimum Credit Ltd's analysts. It is designed to provide
convenience functions, standards, and useful snippets.
Optimum Credit derives significant value from the R
platform and associated community, so non-commercially
sensitive functionality is made available in the spirit of
reciprocity.
}

