% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prob.logit}
\alias{prob.logit}
\title{Convert a probability into a logit}
\usage{
prob.logit(prob)
}
\arguments{
\item{prob}{Probability}
}
\value{
logit Log(odds)
}
\description{
Transforming probabilities into logits (the response from binomial glms)
}
\examples{
prob.logit(0.5) # equals 0
}
\seealso{
\code{\link{prob.odd}}  \code{\link{odd.logit}}

Other creditrisk: \code{\link{giniChart}};
  \code{\link{giniCoef}}; \code{\link{logit.odd}};
  \code{\link{logit.prob}}; \code{\link{odd.logit}};
  \code{\link{odd.prob}}; \code{\link{prob.odd}};
  \code{\link{scaledScore}}
}
\keyword{glm}
\keyword{logit}
\keyword{odds}
\keyword{probability}

