% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sanitise}
\alias{sanitise}
\title{A cleaning function for special characters}
\usage{
sanitise(str)
}
\arguments{
\item{str}{The text to be sanitised}
}
\description{
This function is a helper for cleaning xtable outputs in preperation for PDF production
}
\examples{
sanitise('[&\%#<>\\\\')
}
\seealso{
Other helper: \code{\link{CJ.dt}};
  \code{\link{convertToXML}}; \code{\link{generatePDF}};
  \code{\link{wordwrap}}
}
\keyword{PDF}
\keyword{Rnw}
\keyword{generate}
\keyword{knitr}
\keyword{pdflatex}

