% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wordwrap}
\alias{wordwrap}
\title{Produce a string with one word per line
Designed for splitting strings to fit better on axis on charts}
\usage{
wordwrap(x, ...)
}
\arguments{
\item{x}{string}

\item{...}{Allows additional parameters to be passed to gsub}
}
\description{
Produce a string with one word per line
Designed for splitting strings to fit better on axis on charts
}
\examples{
library('ggplot2')
ggplot(data.frame(x=1:10,y=1:10),aes(x,y))+theme_optimum()+geom_line()
}
\seealso{
Other helper: \code{\link{CJ.dt}};
  \code{\link{convertToXML}}; \code{\link{generatePDF}};
  \code{\link{sanitise}}
}
\keyword{ggplot}

